[EDELIVERY-8239] Fix issue with pMode not showing in the UI

+++ Domibus-MSH/src/main/java/eu/domibus/core/pmode/multitenancy/MultiDomainPModeProvider.java
@@ -284,7 +284,7 @@ public class MultiDomainPModeProvider extends PModeProvider {
     }
 
     @Override
-    public byte[] getPModeFile(int id) {
+    public byte[] getPModeFile(long id) {
         return getCurrentPModeProvider().getPModeFile(id);
     }
